hard code limit 375 nodes
