+++ src/rpc/masternode.cpp
@@ -15,11 +15,13 @@
 #include <masternodeman.h>
 #ifdef ENABLE_WALLET
 #include <privatesend-client.h>
+#include <wallet/coincontrol.h>
 #endif // ENABLE_WALLET
 #include <privatesend-server.h>
 #include <rpc/server.h>
 #include <util.h>
 #include <utilmoneystr.h>
+#include <consensus/validation.h>
 
 #include <fstream>
 #include <iomanip>
@@ -886,6 +888,165 @@ UniValue sentinelping(const JSONRPCRequest& request)
     return true;
 }
 
+/**
+ * @xtdevcoin
+ * this function help user burn correctly their funds to run infinity node
+ */
+static UniValue infinitynodeburnfund(const JSONRPCRequest& request)
+{
+	std::shared_ptr<CWallet> const wallet = GetWalletForJSONRPCRequest(request);
+    CWallet* const pwallet = wallet.get();
+
+    if (!EnsureWalletIsAvailable(pwallet, request.fHelp)) {
+        return NullUniValue;
+    }
+
+    if (request.fHelp || request.params.size() != 1)
+       throw std::runtime_error(
+            "sendtoaddress amount "
+            "\nSend an amount to BurnAddress.\n"
+            "\nArguments:\n"
+            "1. \"amount\"             (numeric or string, required) The amount in " + CURRENCY_UNIT + " to send. eg 0.1\n"
+            "\nResult:\n"
+            "\"BURNtxid\"                  (string) The Burn transaction id. Need to run infinity node\n"
+            "\"CollateralAddress\"         (string) Address of Collateral. Please send 10000 to this address.\n"
+            "\nExamples:\n"
+            + HelpExampleCli("infinitynodeburnfund", "1000000")
+        );
+    // Make sure the results are valid at least up to the most recent block
+    // the user could have gotten from another RPC command prior to now
+    pwallet->BlockUntilSyncedToCurrentChain();
+	
+    if(!masternodeSync.IsMasternodeListSynced())
+	{
+        throw JSONRPCError(RPC_TYPE_ERROR, "Please wait until InfinityNode data is synced!");
+    }
+
+    if (mnodeman.CountSinType(1) >= Params().GetConsensus().nLimitSINNODE_1 ||
+        mnodeman.CountSinType(5) >= Params().GetConsensus().nLimitSINNODE_5 ||
+        mnodeman.CountSinType(10) >= Params().GetConsensus().nLimitSINNODE_10)
+    {
+        throw JSONRPCError(RPC_TYPE_ERROR, "Number of SINNODE is FULL");
+    }
+
+    LOCK2(cs_main, pwallet->cs_wallet);
+    std::string strError;
+    std::vector<COutput> vPossibleCoins;
+    pwallet->AvailableCoins(vPossibleCoins, true, NULL, false, ALL_COINS);
+
+    UniValue results(UniValue::VARR);
+    // Amount
+    CAmount nAmount = AmountFromValue(request.params[0]);
+    if (nAmount != Params().GetConsensus().nMasternodeBurnSINNODE_1 * COIN &&
+        nAmount != Params().GetConsensus().nMasternodeBurnSINNODE_5 * COIN &&
+        nAmount != Params().GetConsensus().nMasternodeBurnSINNODE_10 * COIN)
+    {
+        throw JSONRPCError(RPC_TYPE_ERROR, "Invalid amount to burn and run Infinity node");
+    }
+
+    //Limit node
+    if ((nAmount == Params().GetConsensus().nMasternodeBurnSINNODE_1 * COIN && mnodeman.CountSinType(1) >= Params().GetConsensus().nLimitSINNODE_1) ||
+        (nAmount == Params().GetConsensus().nMasternodeBurnSINNODE_5 * COIN && mnodeman.CountSinType(5) >= Params().GetConsensus().nLimitSINNODE_5) ||
+        (nAmount == Params().GetConsensus().nMasternodeBurnSINNODE_10 * COIN && mnodeman.CountSinType(10) >= Params().GetConsensus().nLimitSINNODE_10) )
+    {
+        strError = strprintf("Error: Number of SINNODE for type %d is FULL", nAmount/COIN);
+        throw JSONRPCError(RPC_TYPE_ERROR, strError);
+    }
+    // BurnAddress
+    CTxDestination dest = DecodeDestination(Params().GetConsensus().cBurnAddress);
+    CScript scriptPubKeyBurnAddress = GetScriptForDestination(dest);
+    std::vector<std::vector<unsigned char> > vSolutions;
+    txnouttype whichType;
+    if (!Solver(scriptPubKeyBurnAddress, whichType, vSolutions))
+        return false;
+    CKeyID keyid = CKeyID(uint160(vSolutions[0]));
+
+	// Wallet comments
+    std::set<CTxDestination> destinations;
+    LOCK(pwallet->cs_wallet);
+    for (COutput& out : vPossibleCoins) {
+        CTxDestination address;
+        const CScript& scriptPubKey = out.tx->tx->vout[out.i].scriptPubKey;
+        bool fValidAddress = ExtractDestination(scriptPubKey, address);
+
+        if (destinations.size() && (!fValidAddress || !destinations.count(address)))
+            continue;
+
+        UniValue entry(UniValue::VOBJ);
+        entry.pushKV("txid", out.tx->GetHash().GetHex());
+        entry.pushKV("vout", out.i);
+
+        if (fValidAddress) {
+            entry.pushKV("address", EncodeDestination(address));
+
+            auto i = pwallet->mapAddressBook.find(address);
+            if (i != pwallet->mapAddressBook.end()) {
+                entry.pushKV("label", i->second.name);
+                if (IsDeprecatedRPCEnabled("accounts")) {
+                    entry.pushKV("account", i->second.name);
+                }
+            }
+
+            if (scriptPubKey.IsPayToScriptHash()) {
+                const CScriptID& hash = boost::get<CScriptID>(address);
+                CScript redeemScript;
+                if (pwallet->GetCScript(hash, redeemScript)) {
+                    entry.pushKV("redeemScript", HexStr(redeemScript.begin(), redeemScript.end()));
+                }
+            }
+        }
+
+        entry.pushKV("scriptPubKey", HexStr(scriptPubKey.begin(), scriptPubKey.end()));
+        entry.pushKV("amount", ValueFromAmount(out.tx->tx->vout[out.i].nValue));
+        entry.pushKV("rawconfirmations", out.nDepth);
+        entry.pushKV("spendable", out.fSpendable);
+        entry.pushKV("solvable", out.fSolvable);
+        entry.pushKV("safe", out.fSafe);
+        if (out.tx->tx->vout[out.i].nValue >= nAmount && out.nDepth >= 2) {
+            // Wallet comments
+            mapValue_t mapValue;
+            bool fSubtractFeeFromAmount = true;
+            bool fUseInstantSend=false;
+            bool fUsePrivateSend=false;
+            CCoinControl coin_control;
+            coin_control.Select(COutPoint(out.tx->GetHash(), out.i));
+
+            CScript script;
+            script = GetScriptForBurn(keyid, "burncoin");
+
+            CReserveKey reservekey(pwallet);
+            CAmount nFeeRequired;
+            CAmount curBalance = pwallet->GetBalance();
+            
+            std::vector<CRecipient> vecSend;
+            int nChangePosRet = -1;
+            CRecipient recipient = {script, nAmount, fSubtractFeeFromAmount};
+            vecSend.push_back(recipient);
+            CTransactionRef tx;
+            if (!pwallet->CreateTransaction(vecSend, tx, reservekey, nFeeRequired, nChangePosRet, strError, coin_control, true, fUsePrivateSend ? ONLY_DENOMINATED : ALL_COINS, fUseInstantSend)) {
+                if (!fSubtractFeeFromAmount && nAmount + nFeeRequired > curBalance)
+                    strError = strprintf("Error: This transaction requires a transaction fee of at least %s", FormatMoney(nFeeRequired));
+                throw JSONRPCError(RPC_WALLET_ERROR, strError);
+            }
+            CValidationState state;
+            if (!pwallet->CommitTransaction(tx, std::move(mapValue), {} /* orderForm */, {}/*fromAccount*/, reservekey, g_connman.get(),
+                            state, fUseInstantSend ? NetMsgType::TXLOCKREQUEST : NetMsgType::TX)) {
+                strError = strprintf("Error: The transaction was rejected! Reason given: %s", FormatStateMessage(state));
+                throw JSONRPCError(RPC_WALLET_ERROR, strError);
+            }
+            entry.pushKV("BURNADDRESS", EncodeDestination(dest));
+            entry.pushKV("BURNPUBLICKEY", HexStr(keyid.begin(), keyid.end()));
+            entry.pushKV("BURNSCRIPT", HexStr(scriptPubKeyBurnAddress.begin(), scriptPubKeyBurnAddress.end()));
+            entry.pushKV("BURNTX", tx->GetHash().GetHex());
+            entry.pushKV("COLLATERAL_ADDRESS",EncodeDestination(address));
+            //coins is good to burn
+            results.push_back(entry);
+            break; //immediat
+        }
+    }
+    return results;
+}
+
 UniValue mnsetup(const JSONRPCRequest& request)
 {
     if (request.fHelp || request.params.size() < 1)
@@ -972,11 +1133,12 @@ UniValue mnsetup(const JSONRPCRequest& request)
 static const CRPCCommand commands[] =
 { //  category              name                      actor (function)         argNames
   //  --------------------- ------------------------  -----------------------  ----------
-    { "dash",               "masternode",             &masternode,             {"command"}  },
+    { "SIN",                "masternode",             &masternode,             {"command"}  },
     { "dash",               "masternodelist",         &masternodelist,         {"mode", "filter"}  },
     { "dash",               "masternodebroadcast",    &masternodebroadcast,    {"command"}  },
     { "dash",               "getpoolinfo",            &getpoolinfo,            {}  },
-    { "dash",               "mnsetup",                &mnsetup,                {}  },
+    { "SIN",                "mnsetup",                &mnsetup,                {}  },
+    { "SIN",                "infinitynodeburnfund",   &infinitynodeburnfund,   {"amount"} },
 #ifdef ENABLE_WALLET
 // SIN TODO:    { "dash",               "privatesend",            &privatesend,            {"command"}  },
 #endif
